refactor(cartesia): move types to where they are used
