+++ livekit-plugins/livekit-plugins-cartesia/livekit/plugins/cartesia/_recognize_streams/auto_finalize_recognize_stream.py
@@ -24,13 +24,11 @@ import aiohttp
 from livekit import rtc
 from livekit.agents import (
     APIConnectionError,
-    APIConnectOptions,
     LanguageCode,
     stt,
     utils,
 )
-from livekit.agents.types import NOT_GIVEN, NotGivenOr
-from livekit.agents.utils.misc import is_given
+from livekit.agents.types import NOT_GIVEN
 
 from ..constants import (
     API_AUTH_HEADER,
@@ -40,11 +38,157 @@ from ..constants import (
     USER_AGENT,
 )
 from ..log import logger
-from ..models import STTEncoding, STTLanguages, TurnDetectingSTTModel
 from .cartesia_recognize_stream import CartesiaRecognizeStream
 
 if TYPE_CHECKING:
-    from .._types.stt_turns_websocket import STTEventMessage
+    from typing import Literal
+
+    from typing_extensions import NotRequired, TypedDict
+
+    from livekit.agents import APIConnectOptions
+    from livekit.agents.types import NotGivenOr
+
+    from ..models import STTEncoding, STTLanguages, TurnDetectingSTTModel
+
+    class STTConnectedEvent(TypedDict):
+        """Fires once when the WebSocket connection is established.
+
+        You do not need to wait for this event before sending audio.
+
+        Attributes:
+            type: Event discriminator.
+            request_id: Unique identifier for this connection. Does not change between turns.
+
+        See also:
+            https://docs.cartesia.ai/api-reference/stt/turns/websocket
+        """
+
+        type: Literal["connected"]
+        request_id: str
+
+    class STTTurnStartEvent(TypedDict):
+        """Model predicts the start of a user turn.
+
+        Attributes:
+            type: Event discriminator.
+            request_id: Unique identifier for this connection. Does not change between turns.
+
+        See also:
+            https://docs.cartesia.ai/api-reference/stt/turns/websocket
+        """
+
+        type: Literal["turn.start"]
+        request_id: str
+
+    class STTTurnUpdateEvent(TypedDict):
+        """Fires repeatedly as the model transcribes the current user turn.
+
+        Used for interim transcript events.
+
+        Attributes:
+            type: Event discriminator.
+            transcript: Cumulative text for the current turn, i.e. the full text transcribed
+                so far in this turn, not a delta.
+            request_id: Unique identifier for this connection. Does not change between turns.
+
+        See also:
+            https://docs.cartesia.ai/api-reference/stt/turns/websocket
+        """
+
+        type: Literal["turn.update"]
+        transcript: str
+        request_id: str
+
+    class STTTurnEagerEndEvent(TypedDict):
+        """Fires when the model predicts the user might be done speaking.
+
+        Used for preflight transcript events.
+
+        Attributes:
+            type: Event discriminator.
+            transcript: Cumulative text for the current turn, i.e. the full text transcribed
+                so far in this turn, not a delta.
+            request_id: Unique identifier for this connection. Does not change between turns.
+
+        See also:
+            https://docs.cartesia.ai/api-reference/stt/turns/websocket
+        """
+
+        type: Literal["turn.eager_end"]
+        transcript: str
+        request_id: str
+
+    class STTTurnResumeEvent(TypedDict):
+        """Fires after ``turn.eager_end`` if the user turn has not actually ended.
+
+        Attributes:
+            type: Event discriminator.
+            request_id: Unique identifier for this connection. Does not change between turns.
+
+        See also:
+            https://docs.cartesia.ai/api-reference/stt/turns/websocket
+        """
+
+        type: Literal["turn.resume"]
+        request_id: str
+
+    class STTTurnEndEvent(TypedDict):
+        """Marks the end of a user turn.
+
+        Used for end-of-speech and final transcript events.
+
+        Attributes:
+            type: Event discriminator.
+            transcript: Cumulative text for the current turn, i.e. the full text transcribed
+                so far in this turn, not a delta.
+            request_id: Unique identifier for this connection. Does not change between turns.
+
+        See also:
+            https://docs.cartesia.ai/api-reference/stt/turns/websocket
+        """
+
+        type: Literal["turn.end"]
+        transcript: str
+        request_id: str
+
+    class STTErrorEvent(TypedDict):
+        """Error event sent by the server.
+
+        Attributes:
+            type: Event discriminator.
+            error_code: Stable code identifying the error.
+            status_code: HTTP-style status code; values >= 500 are treated as retryable.
+            title: Short human-readable error title.
+            message: Detailed human-readable error message.
+            doc_url: URL to documentation describing this error.
+            request_id: Unique identifier for this connection. Does not change between turns.
+
+        See also:
+            https://docs.cartesia.ai/api-reference/stt/turns/websocket
+        """
+
+        type: Literal["error"]
+        error_code: NotRequired[str]
+        status_code: NotRequired[int]
+        title: NotRequired[str]
+        message: NotRequired[str]
+        doc_url: NotRequired[str]
+        request_id: NotRequired[str]
+
+    STTEventMessage = (
+        STTConnectedEvent
+        | STTTurnStartEvent
+        | STTTurnUpdateEvent
+        | STTTurnEagerEndEvent
+        | STTTurnResumeEvent
+        | STTTurnEndEvent
+        | STTErrorEvent
+    )
+    """Server-sent message on the ``/stt/turns/websocket`` endpoint.
+
+    See also:
+        https://docs.cartesia.ai/api-reference/stt/turns/websocket
+    """
 
 
 class AutoFinalizeRecognizeStream(CartesiaRecognizeStream):
@@ -353,7 +497,7 @@ class AutoFinalizeRecognizeStream(CartesiaRecognizeStream):
                 Ink 2 does not have multi-lingual support yet and only works with English.
             model: Deprecated. This is a no-op. Construct a new STT instance to change the model.
         """
-        if is_given(language):
+        if utils.is_given(language):
             self._language = LanguageCode(language)
 
         # not changing the model and reconnecting since that is likely unexpected behavior
